sprites, notes, attacks, and downwards controller
changed sprite rendering, and where images are found. Fix a glitch where the down/shield key would be undefined so a controller player couldn't use it. Made many of the dillion sprites and hooked them up. Fixed a bug where an empty attack function made a cooldown NaN, now defaults to 0 if the cooldown is NaN. Also with attacks you can now see the sprite with each attack. Lessened complexity of some functions and made new functions in the Fighter class.